* May 2026 Canary - TBR 25/05/2026

+++ Scripts/Build/purge.cmd
@@ -8,22 +8,22 @@ if /I "%INPUT%"=="n" goto no
 
 :yes
 echo Deleting the 'Bin' folder
-rd /s /q "..\..\Bin"
+if exist "..\..\Bin\" rd /s /q "..\..\Bin"
 echo Deleted the 'Bin' folder
 echo Deleting the 'Krypton.Docking\obj' folder
-rd /s /q "..\..\Source\Krypton Components\Krypton.Docking\obj"
+if exist "..\..\Source\Krypton Components\Krypton.Docking\obj\" rd /s /q "..\..\Source\Krypton Components\Krypton.Docking\obj"
 echo Deleted the 'Krypton.Docking\obj' folder
 echo Deleting the 'Krypton.Navigator\obj' folder
-rd /s /q "..\..\Source\Krypton Components\Krypton.Navigator\obj"
+if exist "..\..\Source\Krypton Components\Krypton.Navigator\obj\" rd /s /q "..\..\Source\Krypton Components\Krypton.Navigator\obj"
 echo Deleted the 'Krypton.Navigator\obj' folder
 echo Deleting the 'Krypton.Ribbon\obj' folder
-rd /s /q "..\..\Source\Krypton Components\Krypton.Ribbon\obj"
+if exist "..\..\Source\Krypton Components\Krypton.Ribbon\obj\" rd /s /q "..\..\Source\Krypton Components\Krypton.Ribbon\obj"
 echo Deleted the 'Krypton.Ribbon\obj' folder
 echo Deleting the 'Krypton.Toolkit\obj' folder
-rd /s /q "..\..\Source\Krypton Components\Krypton.Toolkit\obj"
+if exist "..\..\Source\Krypton Components\Krypton.Toolkit\obj\" rd /s /q "..\..\Source\Krypton Components\Krypton.Toolkit\obj"
 echo Deleted the 'Krypton.Toolkit\obj' folder
 echo Deleting the 'Krypton.Workspace\obj' folder
-rd /s /q "..\..\Source\Krypton Components\Krypton.Workspace\obj"
+if exist "..\..\Source\Krypton Components\Krypton.Workspace\obj\" rd /s /q "..\..\Source\Krypton Components\Krypton.Workspace\obj"
 echo Deleted the 'Krypton.Workspace\obj' folder
 if exist "..\..\Logs" ( goto deletelogsdirectory ) else echo Directory 'Logs' not found
